Conflicts:
	src/topics/convert_topics.py

+++ data/wikistats/downloading/get_daily_stats.sh
@@ -18,16 +18,25 @@ DATE=`date --date $1 +"%Y%m%d"`
 YEAR=${DATE:0:4}
 MONTH=${DATE:4:2}
 DAY=${DATE:6:2}
-LOGPREFIX="$DATE:"
+LOGPREFIX="$YEAR-$MONTH-$DAY"
+WORKING="working_$DATE"
+LOGFILE="log/daily_stats_$DATE.log"
+
+if [ -e $WORKING ]; then
+	echo "$WORKING directory exists; cancelling..." >&2
+	echo "delete it and try again." >&2
+	exit 1
+fi
 
 mkdir -p log
-rm -rf working_$DATE
-mkdir -p working_$DATE
-cd working_$DATE
+mkdir -p $WORKING
+cd $WORKING
+
+# download the file listing in the latest folder
 wget -nv -o wget.log http://dammit.lt/wikistats/
 if [ -e wget.log ]
 then
-    cat wget.log >> ../log/daily_stats_$DATE.log
+    cat wget.log >> ../$LOGFILE
     rm -f wget.log
 fi
 
@@ -44,37 +53,43 @@ rm -f index.html
 COUNT=0
 if [ $DRYRUN ]; then
     for FILE in $FILES; do
-        echo $FILE
+		BASENAME=`basename $FILE`
+		if [ -e ../../arhive/$YEAR/$MONTH/$BASENAME ]; then
+            echo "$BASENAME already exists; pass" >&2
+		else
+			echo http://dammit.lt/wikistats/$FILE
+			COUNT=$[COUNT+1]
+		fi
     done
 else
-    for FILE in $FILES
-    do
+    for FILE in $FILES; do
         BASENAME=`basename $FILE`
-        if [ -e ../../latest/$YEAR/$MONTH/$BASENAME ]; then
-            echo "$BASENAME already exists; cancel downloading" >&2
+        if [ -e ../../arhive/$YEAR/$MONTH/$BASENAME ]; then
+            echo "$BASENAME already exists; pass" >&2
             continue
         fi
-        mkdir -p ../../latest/$YEAR/$MONTH
+        mkdir -p ../../archive/$YEAR/$MONTH
+		# delete if any file of the same name exists
         rm -f $BASENAME
+		# download the file
         wget -nv -o wget.log http://dammit.lt/wikistats/$FILE
-        if [ -e wget.log ]
-        then
-            cat wget.log >> ../log/daily_stats_$DATE.log
+        if [ -e wget.log ]; then
+            cat wget.log >> ../$LOGFILE
             rm -f wget.log
         fi
-        if [ ! -e $BASENAME ]
-        then
+        if [ ! -e $BASENAME ]; then
             echo "$LOGPREFIX failed to download $FILE" >&2
         else
-            mv $BASENAME ../../latest/$YEAR/$MONTH
-            COUNT=$((COUNT+1))
+            mv $BASENAME ../../archive/$YEAR/$MONTH
+			COUNT=$[COUNT+1]
         fi
     done
 fi
 
+# download the file listing in the archive
 wget -nv -o wget.log http://dammit.lt/wikistats/archive/$YEAR/$MONTH/
 if [ -e wget.log ]; then
-    cat wget.log >> ../log/daily_stats_$DATE.log
+    cat wget.log >> ../$LOGFILE
     rm -f wget.log
 fi
 
@@ -89,77 +104,67 @@ rm -f index.html
 
 if [ $DRYRUN ]; then
     for FILE in $FILES2; do
-        echo $FILE
+		BASENAME=`basename $FILE`
+		if [ -e ../../archive/$YEAR/$MONTH/$BASENAME ]; then
+			echo "$BASENAME already exists; pass" >&2
+			continue
+		else
+			echo http://dammit.lt/wikistats/archive/$YEAR/$MONTH/$FILE
+			COUNT=$[COUNT+1]
+		fi
     done
-    cd ..
-    rmdir working_$DATE
-    exit 0
+else
+	for FILE in $FILES2; do
+		BASENAME=`basename $FILE`
+		if [ -e ../../archive/$YEAR/$MONTH/$BASENAME ]; then
+			echo "$BASENAME already exists; pass" >&2
+			continue
+		fi
+		mkdir -p ../../archive/$YEAR/$MONTH
+		# delete if any file has the same file name
+		rm -f $BASENAME
+		wget -nv -o wget.log http://dammit.lt/wikistats/archive/$YEAR/$MONTH/$FILE
+		if [ -e wget.log ]
+		then
+			cat wget.log >> ../$LOGFILE
+			rm -f wget.log
+		fi
+		if [ ! -e $BASENAME ]
+		then
+			echo "$LOGPREFIX failed to download $FILE" >&2
+		else
+			mv $BASENAME ../../archive/$YEAR/$MONTH
+			COUNT=$[COUNT+1]
+		fi
+	done
 fi
 
-for FILE in $FILES2; do
-    BASENAME=`basename $FILE`
-    if [ -e ../../archive/$YEAR/$MONTH/$BASENAME ]; then
-        echo "$BASENAME already exists; cancel downloading" >&2
-        continue
-    fi
-    mkdir -p ../../archive/$YEAR/$MONTH
-    rm -f $BASENAME
-    wget -nv -o wget.log http://dammit.lt/wikistats/archive/$YEAR/$MONTH/$FILE
-    if [ -e wget.log ]
-    then
-        cat wget.log >> ../log/daily_stats_$DATE.log
-        rm -f wget.log
-    fi
-    if [ ! -e $BASENAME ]
-    then
-        echo "$LOGPREFIX failed to download $FILE" >&2
-    else
-        mv $BASENAME ../../archive/$YEAR/$MONTH
-        COUNT=$((COUNT+1))
-    fi
-done
-
-rm -f SUCCESS
-for FILE in $FILES2; do
-    BASENAME=`basename $FILE`
-    if [ -e ../../archive/$YEAR/$MONTH/$BASENAME -a -e ../../latest/$YEAR/$MONTH/$BASENAME ]; then
-        if diff -q ../../archive/$YEAR/$MONTH/$BASENAME ../../latest/$YEAR/$MONTH/$BASENAME > /dev/null; then
-            echo $BASENAME >> SUCCESS
-        fi
-    fi
-done
-
-if [ -e SUCCESS ]; then
-    for FILE in `cat SUCCESS`; do
-        rm -f ../../latest/$YEAR/$MONTH/$BASENAME
-        COUNT=$((COUNT-1))
-    done
-    rm -f SUCCESS
-fi
 cd ..
-rmdir working_$DATE
+rmdir $WORKING
 
 if [ $COUNT -ne 48 ]
 then
-    echo "$LOGPREFIX missing or redudant: $COUNT files downloaded." >&2
+	echo "$LOGPREFIX files missing or redudant: $COUNT files downloaded." >&2
 fi
 
-for FILE in FILES; do
-    BASENAME=`basename $FILE`
-    if [ -e ../latest/$YEAR/$MONTH/$BASENAME ]; then
-        ../../../src/wikistats/verify_stats.py ../latest/$YEAR/$MONTH/$BASENAME > /dev/null
-        if [ $? -ne 0 ]; then
-            echo "$LOGPREFIX $FILE failed verification." >&2
-        fi
-    fi
-done
+if [ ! $DRYRUN ]; then
+	for FILE in $FILES; do
+		BASENAME=`basename $FILE`
+		if [ -e ../archive/$YEAR/$MONTH/$BASENAME ]; then
+			/mnt/data/wikitopics/src/wikistats/verify_stats.py ../archive/$YEAR/$MONTH/$BASENAME > /dev/null
+			if [ $? -ne 0 ]; then
+				echo "$LOGPREFIX $FILE failed verification." >&2
+			fi
+		fi
+	done
 
-for FILE in FILES2; do
-    BASENAME=`basename $FILE`
-    if [ -e ../archive/$YEAR/$MONTH/$BASENAME ]; then
-        ../../../src/wikistats/verify_stats.py ../archive/$YEAR/$MONTH/$BASENAME > /dev/null
-        if [ $? -ne 0 ]; then
-            echo "$LOGPREFIX $FILE failed verification." >&2
-        fi
-    fi
-done
+	for FILE in $FILES2; do
+		BASENAME=`basename $FILE`
+		if [ -e ../archive/$YEAR/$MONTH/$BASENAME ]; then
+			/mnt/data/wikitopics/src/wikistats/verify_stats.py ../archive/$YEAR/$MONTH/$BASENAME > /dev/null
+			if [ $? -ne 0 ]; then
+				echo "$LOGPREFIX $FILE failed verification." >&2
+			fi
+		fi
+	done
+fi
